Chron

Documentation for Chron.

Chron.BootstrapCrystDistributionKDE

Chron.BootstrapCrystDistributionKDE

Chron.bilinear_exponential

Chron.bilinear_exponential_ll

Chron.check_dist_ll

Chron.metropolis_min

Chron.metropolis_min!

Chron.metropolis_minmax

Chron.metropolis_minmax!

Chron.tMinDistMetropolis

Chron.BootstrapCrystDistributionKDE — Method

BootstrapCrystDistributionKDE(data::AbstractArray, [sigma::AbstractArray])

Bootstrap an estimate of thq;e pre-eruptive (or pre-deposition) mineral crystallization distribution shape from a 2-d array of sample ages (one row per sample, one column per datum, padded with NaNs as needed) and an equivalent-size array of one-sigma uncertainties, using a kernel density estimate of stacked sample data.

Chron.BootstrapCrystDistributionKDE — Method

BootstrapCrystDistributionKDE(smpl::ChronAgeData)

Bootstrap an estimate of the pre-eruptive (or pre-deposition) mineral crystallization distribution shape from a Chron.ChronAgeData object containing data for several samples, using a kernel density estimate of stacked sample data.

Chron.bilinear_exponential — Method

bilinear_exponential(x::Number, p::AbstractVector)

Evaluate the value of a "bilinear exponential" function defined by the parameters p at point x. This function, which can be used to approximate various asymmetric probability distributions found in geochronology, is defined by an exponential function with two log-linear segments joined by an atan sigmoid:

\[ℯ^{p_1 + v x_s p_4^2 p_5^2 - (1-v) x_s p_4^2/p_5^2}\]

where

\[v = 1/2 - atan(x_s)/π\]

is a sigmoid, positive on the left-hand side, and

\[x_s = (x - p_2)/p_3^2\]

is x scaled by mean and variance.

The elements of the parameter array p may be considered to approximately represent

p[1] # pre-exponential (log normaliation constant)
p[2] # mean (central moment)
p[3] # standard deviation
p[4] # sharpness
p[5] # skew

Chron.bilinear_exponential_ll — Method

bilinear_exponential_ll(x, p)

Return the log likelihood corresponding to a bilinear_exponential distribution defined by the parameters p evaluate at point x.

If x is provided as a number and p as a vector, a single evaluation will be returned; if x is provided as a vector and p as a matrix, the sum over all i for each distribution p[:,i] evaluated at x[i] will be returned instead.

See also bilinear_exponential

Chron.check_dist_ll — Method

check_dist_ll(dist::AbstractArray, mu::AbstractArray, sigma::AbstractArray, tmin::Number, tmax::Number)

Return the log-likelihood of a set of mineral ages with means mu and uncertianty sigma being drawn from a given source (i.e., crystallization / closure) distribution dist, with terms to prevent runaway at low N.

Chron.metropolis_min! — Method

metropolis_min!(tminDist::Array, nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

In-place (non-allocating) version of metropolis_min, fills existing array tminDist.

Run a Metropolis sampler to estimate the minimum of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon eruption ages from a distribution of zircon crystallization ages.

Chron.metropolis_min — Method

tminDist = metropolis_min(nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

Run a Metropolis sampler to estimate the minimum of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon eruption ages from a distribution of zircon crystallization ages.

Chron.metropolis_minmax! — Method

metropolis_minmax!(tminDist, tmaxDist, llDist, acceptanceDist, nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

In-place (non-allocating) version of metropolis_minmax, filling existing arrays

Run a Metropolis sampler to estimate the extrema of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon saturation and eruption ages from a distribution of zircon crystallization ages.

Chron.metropolis_minmax — Method

(tminDist, tmaxDist, llDist, acceptanceDist) = metropolis_minmax(nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

Run a Metropolis sampler to estimate the extrema of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon saturation and eruption ages from a distribution of zircon crystallization ages.

Chron.tMinDistMetropolis — Method

smpl = tMinDistMetropolis(smpl::ChronAgeData,nsteps::Int,burnin::Int,dist::Array{Float64})

Calculate the minimum limiting (eruption/deposition) age of each sample in smpl using the metropolis_min function, assuming mineral ages for each sample are drawn from the source distribution dist. Fits a bilinearexponential function to the resulting stationary distribution for each sample.