Chron

Documentation for the Chron.jl package.

Chron.intcal13

Chron.intcal20

Chron.shcal20

Chron.BootstrapCrystDistributionKDE

Chron.BootstrapCrystDistributionKDE

Chron.StratMetropolis

Chron.StratMetropolis14C

Chron.StratMetropolisDist

Chron.bilinear_exponential

Chron.bilinear_exponential_ll

Chron.check_dist_ll

Chron.metropolis_min

Chron.metropolis_min!

Chron.metropolis_minmax

Chron.metropolis_minmax!

Chron.tMinDistMetropolis

Chron.intcal13 — Constant

intcal13
NamedTuple with 4 fields:
  dt            = Int64 1
  Age_Calendar  = StepRange{Int64, Int64}   1:1:50000
  Age_sigma     = Vector{Float64}(50000,)   [8.0 ... 274.0]
  Age_14C       = Vector{Float64}(50000,)   [196.8 ... 46401.0]

The IntCal13 radiocarbon calibration curve

Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PG, Bronk Ramsey C, Buck CE, Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Haflidason H, , Hajdas I, Hatté C, Heaton TJ, Hogg AG, Hughen KA, Kaiser KF, Kromer B, Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Turney CSM, van der Plicht J. (2013) "IntCal13 and MARINE13 radiocarbon age calibration curves 0-50000 years calBP". Radiocarbon 55(4). doi: 10.2458/azujsrc.55.16947

Chron.intcal20 — Constant

intcal20
NamedTuple with 4 fields:
  dt            = Int64 1
  Age_Calendar  = StepRange{Int64, Int64}   1:1:55000
  Age_sigma     = Vector{Float64}(55000,)   [11.0 ... 1024.0]
  Age_14C       = Vector{Float64}(55000,)   [197.0 ... 50100.0]

The IntCal20 northern hemisphere radiocarbon calibration curve.

Reimer P, Austin WEN, Bard E, Bayliss A, Blackwell PG, Bronk Ramsey C, Butzin M, Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Hajdas I, Heaton TJ, Hogg AG, Hughen KA, Kromer B, Manning SW, Muscheler R, Palmer JG, Pearson C, van der Plicht J, Reimer RW, Richards DA, Scott EM, Southon JR, Turney CSM, Wacker L, Adolphi F, Büntgen U, Capano M, Fahrni S, Fogtmann-Schulz A, Friedrich R, Köhler P, Kudsk S, Miyake F, Olsen J, Reinig F, Sakamoto M, Sookdeo A, Talamo S. (2020). "The IntCal20 Northern Hemisphere radiocarbon age calibration curve (0-55 cal kBP)". Radiocarbon 62. doi: 10.1017/RDC.2020.41.

Chron.shcal20 — Constant

shcal20
NamedTuple with 4 fields:
  dt            = Int64 1
  Age_Calendar  = StepRange{Int64, Int64}   1:1:55000
  Age_sigma     = Vector{Float64}(55000,)   [10.0 ... 1008.0]
  Age_14C       = Vector{Float64}(55000,)   [171.0 ... 50132.0]

The SHCal20 southern hemisphere radiocarbon calibration curve.

Hogg AG, Heaton TJ, Hua Q, Palmer JG, Turney CSM, Southon J, Bayliss A, Blackwell PG, Boswijk G, Bronk Ramsey C, Pearson C, Petchey F, Reimer P, Reimer R, Wacker L. (2020). "SHCal20 Southern Hemisphere calibration, 0-55,000 years cal BP". Radiocarbon 62. doi: 10.1017/RDC.2020.59

Chron.BootstrapCrystDistributionKDE — Method

BootstrapCrystDistributionKDE(data::AbstractArray, [sigma::AbstractArray]; cutoff=-0.05)

Bootstrap an estimate of the pre-eruptive (or pre-depositional) mineral crystallization distribution shape from a 1- or 2-d array of sample ages (one row per sample, one column per datum, padded with NaNs as needed) and an equivalent-size array of one-sigma uncertainties, using a kernel density estimate of stacked sample data.

Examples

# Bootstrap crystallization distribution for a synthetic dataset with ages
# [1,2,3,...10] Ma and uncertainties of 1 Ma each
BootstrappedDistribution = BootstrapCrystDistributionKDE(1:10, ones(10))

Chron.BootstrapCrystDistributionKDE — Method

BootstrapCrystDistributionKDE(smpl::ChronAgeData; cutoff=-0.05)

Bootstrap an estimate of the pre-eruptive (or pre-depositional) mineral crystallization distribution shape from a Chron.ChronAgeData object containing data for several samples, using a kernel density estimate of stacked sample data.

Examples

BootstrappedDistribution = BootstrapCrystDistributionKDE(smpl)

Chron.StratMetropolis — Method

StratMetropolis(smpl::ChronAgeData, [hiatus::HiatusData,] config::StratAgeModelConfiguration)

Runs the main Chron.jl age-depth model routine for a stratigraphic set of samples defined by sample heights and simple Gaussian age constraints in the smpl struct, and an age-depth model configuration defined by the config struct.

Optionally, if a hiatus struct is provided, the model will additionally incorporate information about the durations of known hiatuses at the specified model heights.

Examples:

(mdl, agedist, lldist) = StratMetropolis(smpl, config)

(mdl, agedist, hiatusdist, lldist) = StratMetropolis(smpl, hiatus, config)

Chron.StratMetropolis14C — Method

StratMetropolis14C(smpl::ChronAgeData, [hiatus::HiatusData,] config::StratAgeModelConfiguration)

Runs the main Chron.jl age-depth model routine for a stratigraphic set of samples defined by sample heights and interpolated radiocarbon age constraints in the smpl struct, and an age-depth model configuration defined by the config struct.

Optionally, if a hiatus struct is provided, the model will additionally incorporate information about the durations of known hiatuses at the specified model heights.

Examples:

(mdl, agedist, lldist) = StratMetropolis14C(smpl, config)

(mdl, agedist, hiatusdist, lldist) = StratMetropolis14C(smpl, hiatus, config)

Chron.StratMetropolisDist — Method

StratMetropolisDist(smpl::ChronAgeData, [hiatus::HiatusData,] config::StratAgeModelConfiguration)

Runs the main Chron.jl age-depth model routine for a stratigraphic set of samples defined by sample heights and fitted asymmetric age distributions (bilinear_exponential) in the smpl struct, and an age-depth model configuration defined by the config struct.

Optionally, if a hiatus struct is provided, the model will additionally incorporate information about the durations of known hiatuses at the specified model heights.

Examples:

(mdl, agedist, lldist) = StratMetropolisDist(smpl, config)

(mdl, agedist, hiatusdist, lldist) = StratMetropolisDist(smpl, hiatus, config)

Chron.bilinear_exponential — Method

bilinear_exponential(x::Number, p::AbstractVector)

Evaluate the value of a "bilinear exponential" function defined by the parameters p at point x. This function, which can be used to approximate various asymmetric probability distributions found in geochronology, is defined by an exponential function with two log-linear segments joined by an atan sigmoid:

\[ℯ^{p_1 + v x_s p_4^2 p_5^2 - (1-v) x_s p_4^2/p_5^2}\]

where

\[v = 1/2 - atan(x_s)/π\]

is a sigmoid, positive on the left-hand side, and

\[x_s = (x - p_2)/p_3^2\]

is x scaled by mean and variance.

The elements of the parameter array p may be considered to approximately represent

p[1] # pre-exponential (log normaliation constant)
p[2] # mean (central moment)
p[3] # standard deviation
p[4] # sharpness
p[5] # skew

Chron.bilinear_exponential_ll — Method

bilinear_exponential_ll(x, p)

Return the log likelihood corresponding to a bilinear_exponential distribution defined by the parameters p evaluate at point x.

If x is provided as a number and p as a vector, a single evaluation will be returned; if x is provided as a vector and p as a matrix, the sum over all i for each distribution p[:,i] evaluated at x[i] will be returned instead.

See also bilinear_exponential

Chron.check_dist_ll — Method

check_dist_ll(dist::AbstractArray, mu::AbstractArray, sigma::AbstractArray, tmin::Number, tmax::Number)

Return the log-likelihood of a set of mineral ages with means mu and uncertianty sigma being drawn from a given source (i.e., crystallization / closure) distribution dist, with terms to prevent runaway at low N.

Examples

mu, sigma = collect(100:0.1:101), 0.01*ones(11)
ll = check_dist_ll(MeltsVolcanicZirconDistribution, mu, sigma, 100, 101)

Chron.metropolis_min! — Method

metropolis_min!(tminDist::Array, nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

In-place (non-allocating) version of metropolis_min, fills existing array tminDist.

Run a Metropolis sampler to estimate the minimum of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon eruption ages from a distribution of zircon crystallization ages.

Examples

metropolis_min!(tminDist, 2*10^5, MeltsVolcanicZirconDistribution, mu, sigma, burnin=10^5)

Chron.metropolis_min — Method

tminDist = metropolis_min(nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

Run a Metropolis sampler to estimate the minimum of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon eruption ages from a distribution of zircon crystallization ages.

Examples

tmindist = metropolis_min(2*10^5, MeltsVolcanicZirconDistribution, mu, sigma, burnin=10^5)

Chron.metropolis_minmax! — Method

metropolis_minmax!(tminDist, tmaxDist, llDist, acceptanceDist, nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

In-place (non-allocating) version of metropolis_minmax, filling existing arrays

Run a Metropolis sampler to estimate the extrema of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon saturation and eruption ages from a distribution of zircon crystallization ages.

Examples

metropolis_minmax!(tmindist, tmaxdist, lldist, acceptancedist, 2*10^5, MeltsVolcanicZirconDistribution, mu, sigma, burnin=10^5)

Chron.metropolis_minmax — Method

metropolis_minmax(nsteps::Int, dist::AbstractArray, data::AbstractArray, uncert::AbstractArray; burnin::Integer=0)

Run a Metropolis sampler to estimate the extrema of a finite-range source distribution dist using samples drawn from that distribution – e.g., estimate zircon saturation and eruption ages from a distribution of zircon crystallization ages.

Examples

tmindist, tmaxdist, lldist, acceptancedist = metropolis_minmax(2*10^5, MeltsVolcanicZirconDistribution, mu, sigma, burnin=10^5)

Chron.tMinDistMetropolis — Method

tMinDistMetropolis(smpl::ChronAgeData, nsteps::Int, burnin::Int, dist::Array{Float64})

Calculate the minimum limiting (eruption/deposition) age of each sample defined in the smpl struct, using the metropolis_min function, assuming mineral ages for each sample are drawn from the source distribution dist. Fits a bilinear_exponential function to the resulting stationary distribution for each sample and stores the results in smpl.Params for use by the StratMetropolisDist function.

Examples

smpl = tMinDistMetropolis(smpl, 5*10^5, 2*10^5, TriangularDistribution)